file_manager/requirements.txt
Django==2.2.3
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
